moving gamma max and gamma break functions to a separate file
